Skiller ut slange-funksjonalitet i useSnake
