+++ src/components/games/SpareslagenSpill/snake/useDirection.ts
